light/dark mode toggle enabled. fixed review details page. added gameslistpage for list of a category of games. added tests to create reviews and comments for reviews

+++ Backend/Models/DTO/Social/ReviewDto.cs
@@ -11,8 +11,8 @@ namespace Backend.Models.DTO.Social
         public Guid GameId { get; set; }
 
         [Required]
-        [Range(1, 5, ErrorMessage = "Rating must be between 1 and 5")]
-        public int Rating { get; set; }
+        [Range(1.0, 5.0, ErrorMessage = "Rating must be between 1 and 5")]
+        public double Rating { get; set; }
 
         [StringLength(1000, ErrorMessage = "Review text cannot exceed 1000 characters")]
         public string? ReviewText { get; set; }
